Tokenizer is splitting away from main
desu. Still does some things that indexer should be doing... It'll take
a while, but it's possible to have the tokenizer ONLY spit out strings,
and do the actual indexing in indexer. I'd have to make a struct
implementation, and create the struct in Indexer, and split the deleting
off from the merge function too, call that at the end. Also would
require regexing the entire file at once? Consider.